Added TODO files for "Accounts", new project based on accounts-sso
With this commit we start a new project called "Accounts", which is
going to be developed alongside WebAccounts so we can keep history of
the PoC we have done, and at the same time be sure that we are feature
equivalent once we finish with the new project.

Also, we are going to need some of the code (specially the one that
gets accounts from KConfig) to do an importer from old WebAccounts
to the new accounts-sso framework.