First code drop of Accounts KDED
This KDED will be in charge of enabling/disabling KDE services such
akonadi, perhaps Telapathy etc.

One of the things it will do (for example) is to create a PlavesView
when an ownCloud account is configured.

ATM it kinda works for Akonadi though I'm pushing this so mck182 can do
some testing.