Look for the serviceType instead of service Name
This makes it waaaay easier to specify supported services in the
resources since instead of having to add one mimetype for service it
will be enough to add one with the serviceType. For example:

text/x-vnd.accounts.twitter-microblog
or
text/x-vnd.accounts.imap-email

Using mimetypes for this is still hackish but it will do for now.