Add a job to check if an owncloud server is correct
This job will look for the right url so it is a little bit prepared for
typos or different ways of introducing an url, for example:

http://owncloud.com
http://owncloud.com/
https://owncloud.com
owncloud.com
owncloud.com/dav/files

All that are correct ways of indicating the server to check, the idea is
to work on all possible ways an user can input the host.

The check is done by getting the file server/status.php and checking
that there is a version, good enough for what we want.