Add KAccounts support library
This allows to extend KAccounts functionality with 3rd party plugins
like KDE Telepathy plugins.

The library provides two interfaces - KAccountsDPlugin and
KAccountsUiPlugin.

KAccountsDPlugin is the interface for the daemon part of KAccounts; this
reacts to accounts being created or removed.

KAccountsUiPlugin is the interface for showing a specialized UIs while
creating new accounts. This is currently used for KTp accounts - the
plugin is build on top of ktp-accounts-kcm internal libraries. This
nicely moves all the Telepathy interaction where it belongs.