Handle creating KTp accounts using the KTp UI plugin
This loads the KTp UI plugin using QPluginLoader, instantiates the
plugin and shows the KTp dialog for creating new accounts. The dialog
then returns the username, password and any additional data. Username
and password are stored in the SSO storage and since there's no way to
set the Accounts-SSO account id on the Telepathy account, it's stored
into a config file, where ktp-auth-handler then reads it from.