Disable create form while creating new accounts
When adding a new account in the KCM, it can take a while from the click
on the provider's button to the moment the window is presented to the user.
For example this happens with Google accounts, since we need to wait for
webkit plus the authentication page from Google.

To be sure that the users don't click twice the buttons while they are
waiting, disable the whole form and re-enable it later when the
authentication is done.

Reviewers: mck182

Differential Revision: https://phabricator.kde.org/D5727