Add error message if intltool-merge is not found
Users of the kaccounts_add_* macros are supposed to do a
find_package(Intltool REQUIRED), but if they don't the build fails with
a rather obscure message (`make install` fails because the service or
provider file was not generated).

This error message will at least make the life of packagers a bit
easier.

BUG: 378932

Differential Revision: https://phabricator.kde.org/D6114