Unbreak the Accounts KCM
Summary:
Prior to this, the Accounts KCM would show only the first account,
because the listview was a zero by zero size (which results in
the first item still being shown because labels will overpaint
unless told otherwise). The reson it looked like there was a
view there before was that the view was set to the ScrollViewKCM's
default scroll view. In other words: All the accounts were in fact
there, but they were not displayed because the listview was
not there.

BUG:415267

Reviewers: #plasma, nicolasfella, bshah

Reviewed By: #plasma, bshah

Subscribers: arojas

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27454