Polish up the KAccounts KCM
Summary:
Since changing to the QtQuick based KAccounts KCM, it seems that
we managed to lose a few features that really want to be in there.
This is work toward ensuring that we get those features back, and
further end up with something which is more pleasant in use in
general than what we had before.

* Add a job for toggling an account service's enabled status and use it to allow this (and then force the model to update, because apparently it doesn't do that automatically)
* Add a dialog when deleting accounts (to ensure less accidental data loss when just clicking around the UI)
* Add self to copyright
* Various little bits of polish (like swapping out icons, ensuring wording is more pleasant in a few places, and adding tooltips)

Accounts list with a few accounts (hovering over one):
{F8112625}

Clicked the remove account action for an account:
{F8112366}

Reviewers: #plasma, bshah, nicolasfella

Reviewed By: nicolasfella

Subscribers: nicolasfella, ahiemstra, apol, mart, bshah, jbbgameich

Differential Revision: https://phabricator.kde.org/D27479