[CreateAccount job] Never set an empty name when creating an account
At least when creating a google account, signon-ui isn't able to extract
the username from the login page anymore since a couple of months (and
newer versions ported to QtWebEngine don't even extract it at all
anymore because it was considered to be too fragile, which apparently
proved to be true...).

This looks kind of broken, and actually made kio-gdrive useless as that
doesn't even list accounts with an empty name.

To fix it, set the name to some arbitrary string in the case that
info.userName() is empty.

BUG: 414219
FIXED-IN: 19.12.3
Differential Revision: https://phabricator.kde.org/D27474