Port query logic for ui plugins to KPluginMetadata::findPlugins
With https://phabricator.kde.org/T15094 we can allow plugins without
embedded json metadata to be consiered valid and consequently
show up in the KPluginMetaData::findPlugins result.

This way we can drop the custom logic to iterate over the library paths
and check the containing files.

Also we can utilize the pluginId that is derived from the filename
instead of meddling with the filename manually and appending the file
extension when accessing our internal hash of loaded plugins.